moved main script to root folder, changed workflow
