+++ .github/workflows/build.yml
@@ -4,90 +4,44 @@ on: [push, pull_request]
 
 jobs:
     build:
-        name: Build binaries [${{ matrix.os }}-${{ matrix.python-version}}]
+        name: Build binaries
+        
         strategy:
             matrix:
-                include:
-                    - os: macos-14
-                    - os: windows-2019
-                    - os: ubuntu-20.04
-                python-version: ['3.11']
+                os: [macos-latest, ubuntu-latest, windows-latest]
+        
         runs-on: "${{ matrix.os }}"
+        
         steps:
             - name: Checkout repository
               uses: actions/checkout@v4
 
-            - name: Setup python [${{ matrix.python-version }}]
-              uses: actions/setup-python@v4
+            - name: Setup python
+              uses: actions/setup-python@v5
               with:
-                python-version: ${{ matrix.python-version }}
-
-            - name: Install poetry
-              uses: Gr1N/setup-poetry@v8
+                python-version: '3.11'
+                cache: 'pip'
+                cache-dependency-path: |
+                  **/requirements*.txt
 
             - name: Install dependencies
               run: |
-                poetry run pip install --upgrade pip setuptools
-                poetry install
+                pip install -r requirements.txt
 
-            - name: Build python releases
-              run: |
-                poetry build
-
-            - name: Build executable => Linux
-              if: ${{ !startsWith(matrix.os, 'windows') }}
-              run: |
-                poetry run pip install nuitka
-                poetry run python -m nuitka \
-                  --standalone \
-                  --onefile \
-                  src/main.py
-
-            - name: Build executable => Windows
-              if: ${{ startsWith(matrix.os, 'windows') }}
-              run: |
-                poetry run pip install nuitka
-                poetry run python -m nuitka `
-                  --standalone `
-                  --onefile `
-                  src/main.py
+            - name: Build executable
+              uses: Nuitka/Nuitka-Action@main
+              with:
+                nuitka-version: main
+                script-name: MMR-Fido.py
+                mode: app
+                enable-plugins: pyqt6
             
-            - name: Upload builds
-              uses: actions/upload-artifact@v3
+            - name: Upload artifacts
+              uses: actions/upload-artifact@v4
               with:
-                name: built-binary
+                name: ${{ runner.os }} Build
                 path: |
-                  dist/*.whl
-                  dist/*.tar.gz
                   build/*.exe
                   build/*.bin
-
-    finalize:
-        name: Finalize
-        runs-on: ubuntu-latest
-        needs: build
-        permissions:
-            contents: write
-        steps:
-            - name: Checkout repository
-              uses: actions/checkout@v4
-
-            - name: Download artifacts
-              uses: actions/download-artifact@v3
-              with:
-                name: built-binary
-
-            - name: Create release
-              id: create_release
-              uses: softprops/actions-gh-release@v1
-              env:
-                GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-              with:
-                tag_name: ${{ github.ref }}
-                name: Release ${{ github.ref_name }}
-                draft: false
-                files: |
-                  dist/*.whl
-                  dist/*.tar.gz
-                  build/*.exe
-                  build/*.bin
+                  build/*.app/**/*
+                include-hidden-files: true
