+++ requirements.txt
@@ -1,0 +1,7 @@
+contourpy==1.3.1
+matplotlib==3.9.3
+nmrglue==0.10
+numpy==2.1.3
+PyQt6==6.7.1
+PyQt6_sip==13.8.0
+pyqtgraph==0.13.7
